// src/app/hpp/models/FormValues.ts
export interface FormValues {
  nilaiTukarUSD: number;
  hargaSapiCIF: number;
  beratSapi: number;
  beaMasuk: number;
  pph: number;
  biayaImport: number;
  lamaPemeliharaan: number;
  biayaPemeliharaan: number;
  pertambahanBerat: number;
  tingkatBungaBank: number;
  biayaOverHead: number;
  biayaMarketing: number;
  hargaJualSapi: number;
}